Move ApplyVersions from NetKAN to Core
