+++ Netkan/Model/Metadata.cs
@@ -7,7 +7,6 @@ using YamlDotNet.RepresentationModel;
 using CKAN.Versioning;
 using CKAN.Extensions;
 using CKAN.NetKAN.Transformers;
-using CKAN.NetKAN.Services;
 
 namespace CKAN.NetKAN.Model
 {
@@ -180,7 +179,7 @@ namespace CKAN.NetKAN.Model
                 first.Merge(other, mergeSettings);
             }
             // Merge game version compatibility
-            ModuleService.ApplyVersions(first, null, null, null);
+            GameVersion.SetJsonCompatibility(first, null, null, null);
             // Only generate array if multiple URLs
             first[DownloadPropertyName] = downloads.Length == 1
                 ? downloads.First()
