+++ Netkan/Transformers/AvcTransformer.cs
@@ -12,6 +12,7 @@ using CKAN.NetKAN.Validators;
 using CKAN.Avc;
 using CKAN.Games;
 using CKAN.NetKAN.Sources.Github;
+using CKAN.Versioning;
 
 namespace CKAN.NetKAN.Transformers
 {
@@ -140,7 +141,7 @@ namespace CKAN.NetKAN.Transformers
 
         public static void ApplyVersions(JObject json, AvcVersion avc)
         {
-            ModuleService.ApplyVersions(json, avc.ksp_version, avc.ksp_version_min, avc.ksp_version_max);
+            GameVersion.SetJsonCompatibility(json, avc.ksp_version, avc.ksp_version_min, avc.ksp_version_max);
 
             if (avc.version != null)
             {
